Reexporting `Point`, `Rect` and `Size` at top level
