+++ src/lib.rs
@@ -33,6 +33,10 @@ pub use clip_draw_state::clip_draw_state;
 pub use inside_draw_state::inside_draw_state;
 pub use outside_draw_state::outside_draw_state;
 
+pub use point::Point;
+pub use rect::Rect;
+pub use size::Size;
+
 pub use context::Context;
 
 /// Any triangulation method called on the back-end
@@ -50,6 +54,10 @@ mod clip_draw_state;
 mod inside_draw_state;
 mod outside_draw_state;
 
+mod point;
+mod rect;
+mod size;
+
 pub mod character;
 pub mod context;
 pub mod color;
@@ -59,7 +67,6 @@ pub mod ellipse;
 pub mod rectangle;
 pub mod image;
 pub mod types;
-pub mod types_new;
 pub mod modular_index;
 pub mod text;
 pub mod triangulation;
@@ -72,10 +79,6 @@ pub mod radians {
 
     pub use vecmath_lib::consts::Radians;
 }
-
-/// A rectangle property
-#[derive(Copy, Clone)]
-pub struct Rect(pub types::Rectangle);
 
 /// Clears the screen.
 pub fn clear<G>(
